Revert "Add KubernetesPodOperator & DockerOperator management (#133)"
This reverts commit af87315d53080d7e5fc8c9155a1c1c3fb4040970.
