+++ cosmos/providers/dbt/__init__.py
@@ -2,53 +2,28 @@
 dbt support for Airflow. Contains dags, task groups, and operators.
 """
 
-from cosmos.providers.dbt.core.operators.docker import (
-    DbtLSDockerOperator,
-    DbtRunDockerOperator,
-    DbtRunOperationDockerOperator,
-    DbtSeedDockerOperator,
-    DbtTestDockerOperator,
-)
-from cosmos.providers.dbt.core.operators.kubernetes import (
-    DbtLSKubernetesOperator,
-    DbtRunKubernetesOperator,
-    DbtRunOperationKubernetesOperator,
-    DbtSeedKubernetesOperator,
-    DbtTestKubernetesOperator,
-)
-
 # re-export the operators
-from cosmos.providers.dbt.core.operators.local import (
-    DbtLSLocalOperator,
-    DbtRunLocalOperator,
-    DbtRunOperationLocalOperator,
-    DbtSeedLocalOperator,
-    DbtTestLocalOperator,
+from .core.operators import (
+    DbtLSOperator,
+    DbtRunOperationOperator,
+    DbtRunOperator,
+    DbtSeedOperator,
+    DbtTestOperator,
 )
 
 # re-export user facing utilities
-from cosmos.providers.dbt.core.utils.data_aware_scheduling import get_dbt_dataset
+from .core.utils.data_aware_scheduling import get_dbt_dataset
 
 # re-export the dag and task group
-from cosmos.providers.dbt.dag import DbtDag
-from cosmos.providers.dbt.task_group import DbtTaskGroup
+from .dag import DbtDag
+from .task_group import DbtTaskGroup
 
 __all__ = [
-    DbtLSLocalOperator,
-    DbtRunOperationLocalOperator,
-    DbtRunLocalOperator,
-    DbtSeedLocalOperator,
-    DbtTestLocalOperator,
-    DbtLSDockerOperator,
-    DbtRunOperationDockerOperator,
-    DbtRunDockerOperator,
-    DbtSeedDockerOperator,
-    DbtTestDockerOperator,
-    DbtLSKubernetesOperator,
-    DbtRunOperationKubernetesOperator,
-    DbtRunKubernetesOperator,
-    DbtSeedKubernetesOperator,
-    DbtTestKubernetesOperator,
+    DbtLSOperator,
+    DbtRunOperationOperator,
+    DbtRunOperator,
+    DbtSeedOperator,
+    DbtTestOperator,
     get_dbt_dataset,
     DbtDag,
     DbtTaskGroup,
