--- docs/autoapi/cosmos/core/graph/index.rst
@@ -1,13 +1,0 @@
-:py:mod:`cosmos.core.graph`
-===========================
-
-.. py:module:: cosmos.core.graph
-
-
-Submodules
-----------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 1
-
-   entities/index.rst
