--- docs/autoapi/cosmos/core/index.rst
@@ -1,22 +1,0 @@
-:py:mod:`cosmos.core`
-=====================
-
-.. py:module:: cosmos.core
-
-
-Subpackages
------------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 3
-
-   graph/index.rst
-
-
-Submodules
-----------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 1
-
-   airflow/index.rst
