--- docs/autoapi/cosmos/index.rst
@@ -1,26 +1,0 @@
-:py:mod:`cosmos`
-================
-
-.. py:module:: cosmos
-
-.. autoapi-nested-parse::
-
-   Astronomer Cosmos is a library for rendering 3rd party workflows in Airflow.
-
-
-
-Subpackages
------------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 3
-
-   core/index.rst
-   providers/index.rst
-
-
-Package Contents
-----------------
-
-.. py:data:: __version__
-   :value: '0.4.3dev'
