--- docs/autoapi/cosmos/providers/dbt/constants/index.rst
@@ -1,10 +1,0 @@
-:py:mod:`cosmos.providers.dbt.constants`
-========================================
-
-.. py:module:: cosmos.providers.dbt.constants
-
-
-Module Contents
----------------
-
-.. py:data:: DBT_PROFILE_PATH
