--- docs/autoapi/cosmos/providers/dbt/core/index.rst
@@ -1,15 +1,0 @@
-:py:mod:`cosmos.providers.dbt.core`
-===================================
-
-.. py:module:: cosmos.providers.dbt.core
-
-
-Subpackages
------------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 3
-
-   operators/index.rst
-   profiles/index.rst
-   utils/index.rst
