--- docs/autoapi/cosmos/providers/dbt/core/operators/index.rst
@@ -1,16 +1,0 @@
-:py:mod:`cosmos.providers.dbt.core.operators`
-=============================================
-
-.. py:module:: cosmos.providers.dbt.core.operators
-
-
-Submodules
-----------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 1
-
-   base/index.rst
-   docker/index.rst
-   kubernetes/index.rst
-   local/index.rst
