--- docs/autoapi/cosmos/providers/dbt/core/utils/data_aware_scheduling/index.rst
@@ -1,20 +1,0 @@
-:py:mod:`cosmos.providers.dbt.core.utils.data_aware_scheduling`
-===============================================================
-
-.. py:module:: cosmos.providers.dbt.core.utils.data_aware_scheduling
-
-
-Module Contents
----------------
-
-
-Functions
-~~~~~~~~~
-
-.. autoapisummary::
-
-   cosmos.providers.dbt.core.utils.data_aware_scheduling.get_dbt_dataset
-
-
-
-.. py:function:: get_dbt_dataset(connection_id: str, project_name: str, model_name: str)
