--- docs/autoapi/cosmos/providers/dbt/core/utils/index.rst
@@ -1,15 +1,0 @@
-:py:mod:`cosmos.providers.dbt.core.utils`
-=========================================
-
-.. py:module:: cosmos.providers.dbt.core.utils
-
-
-Submodules
-----------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 1
-
-   data_aware_scheduling/index.rst
-   file_syncing/index.rst
-   profiles_generator/index.rst
