--- docs/autoapi/cosmos/providers/dbt/index.rst
@@ -1,31 +1,0 @@
-:py:mod:`cosmos.providers.dbt`
-==============================
-
-.. py:module:: cosmos.providers.dbt
-
-.. autoapi-nested-parse::
-
-   dbt support for Airflow. Contains dags, task groups, and operators.
-
-
-
-Subpackages
------------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 3
-
-   core/index.rst
-   parser/index.rst
-
-
-Submodules
-----------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 1
-
-   constants/index.rst
-   dag/index.rst
-   render/index.rst
-   task_group/index.rst
