--- docs/autoapi/cosmos/providers/dbt/parser/index.rst
@@ -1,13 +1,0 @@
-:py:mod:`cosmos.providers.dbt.parser`
-=====================================
-
-.. py:module:: cosmos.providers.dbt.parser
-
-
-Submodules
-----------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 1
-
-   project/index.rst
