--- docs/autoapi/cosmos/providers/index.rst
@@ -1,13 +1,0 @@
-:py:mod:`cosmos.providers`
-==========================
-
-.. py:module:: cosmos.providers
-
-
-Subpackages
------------
-.. toctree::
-   :titlesonly:
-   :maxdepth: 3
-
-   dbt/index.rst
