--- docs/autoapi/index.rst
@@ -1,11 +1,0 @@
-API Reference
-=============
-
-This page contains auto-generated API reference documentation [#f1]_.
-
-.. toctree::
-   :titlesonly:
-
-   /autoapi/cosmos/index
-
-.. [#f1] Created with `sphinx-autoapi <https://github.com/readthedocs/sphinx-autoapi>`_
