+++ docs/reference/_generated/cosmos.providers.dbt.DbtDag.rst
@@ -1,0 +1,16 @@
+﻿cosmos.providers.dbt.DbtDag
+===========================
+
+.. currentmodule:: cosmos.providers.dbt
+
+.. autoclass:: DbtDag
+
+
+   .. automethod:: __init__
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+
+      ~DbtDag.__init__
