+++ docs/reference/_generated/cosmos.providers.dbt.DbtLSOperator.rst
@@ -1,0 +1,36 @@
+﻿cosmos.providers.dbt.DbtLSOperator
+==================================
+
+.. currentmodule:: cosmos.providers.dbt
+
+.. autoclass:: DbtLSOperator
+
+
+   .. automethod:: __init__
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+
+      ~DbtLSOperator.__init__
+      ~DbtLSOperator.add_global_flags
+      ~DbtLSOperator.build_and_run_cmd
+      ~DbtLSOperator.exception_handling
+      ~DbtLSOperator.execute
+      ~DbtLSOperator.get_dbt_path
+      ~DbtLSOperator.get_env
+      ~DbtLSOperator.on_kill
+      ~DbtLSOperator.run_command
+      ~DbtLSOperator.subprocess_hook
+
+
+
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~DbtLSOperator.template_fields
+      ~DbtLSOperator.ui_color
