+++ docs/reference/_generated/cosmos.providers.dbt.DbtRunOperationOperator.rst
@@ -1,0 +1,37 @@
+﻿cosmos.providers.dbt.DbtRunOperationOperator
+============================================
+
+.. currentmodule:: cosmos.providers.dbt
+
+.. autoclass:: DbtRunOperationOperator
+
+
+   .. automethod:: __init__
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+
+      ~DbtRunOperationOperator.__init__
+      ~DbtRunOperationOperator.add_cmd_flags
+      ~DbtRunOperationOperator.add_global_flags
+      ~DbtRunOperationOperator.build_and_run_cmd
+      ~DbtRunOperationOperator.exception_handling
+      ~DbtRunOperationOperator.execute
+      ~DbtRunOperationOperator.get_dbt_path
+      ~DbtRunOperationOperator.get_env
+      ~DbtRunOperationOperator.on_kill
+      ~DbtRunOperationOperator.run_command
+      ~DbtRunOperationOperator.subprocess_hook
+
+
+
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~DbtRunOperationOperator.template_fields
+      ~DbtRunOperationOperator.ui_color
