+++ docs/reference/_generated/cosmos.providers.dbt.DbtRunOperator.rst
@@ -1,0 +1,37 @@
+﻿cosmos.providers.dbt.DbtRunOperator
+===================================
+
+.. currentmodule:: cosmos.providers.dbt
+
+.. autoclass:: DbtRunOperator
+
+
+   .. automethod:: __init__
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+
+      ~DbtRunOperator.__init__
+      ~DbtRunOperator.add_global_flags
+      ~DbtRunOperator.build_and_run_cmd
+      ~DbtRunOperator.exception_handling
+      ~DbtRunOperator.execute
+      ~DbtRunOperator.get_dbt_path
+      ~DbtRunOperator.get_env
+      ~DbtRunOperator.on_kill
+      ~DbtRunOperator.run_command
+      ~DbtRunOperator.subprocess_hook
+
+
+
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~DbtRunOperator.template_fields
+      ~DbtRunOperator.ui_color
+      ~DbtRunOperator.ui_fgcolor
