+++ docs/reference/_generated/cosmos.providers.dbt.DbtSeedOperator.rst
@@ -1,0 +1,37 @@
+﻿cosmos.providers.dbt.DbtSeedOperator
+====================================
+
+.. currentmodule:: cosmos.providers.dbt
+
+.. autoclass:: DbtSeedOperator
+
+
+   .. automethod:: __init__
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+
+      ~DbtSeedOperator.__init__
+      ~DbtSeedOperator.add_cmd_flags
+      ~DbtSeedOperator.add_global_flags
+      ~DbtSeedOperator.build_and_run_cmd
+      ~DbtSeedOperator.exception_handling
+      ~DbtSeedOperator.execute
+      ~DbtSeedOperator.get_dbt_path
+      ~DbtSeedOperator.get_env
+      ~DbtSeedOperator.on_kill
+      ~DbtSeedOperator.run_command
+      ~DbtSeedOperator.subprocess_hook
+
+
+
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~DbtSeedOperator.template_fields
+      ~DbtSeedOperator.ui_color
