+++ docs/reference/_generated/cosmos.providers.dbt.DbtTaskGroup.rst
@@ -1,0 +1,16 @@
+﻿cosmos.providers.dbt.DbtTaskGroup
+=================================
+
+.. currentmodule:: cosmos.providers.dbt
+
+.. autoclass:: DbtTaskGroup
+
+
+   .. automethod:: __init__
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+
+      ~DbtTaskGroup.__init__
