+++ docs/reference/_generated/cosmos.providers.dbt.DbtTestOperator.rst
@@ -1,0 +1,36 @@
+﻿cosmos.providers.dbt.DbtTestOperator
+====================================
+
+.. currentmodule:: cosmos.providers.dbt
+
+.. autoclass:: DbtTestOperator
+
+
+   .. automethod:: __init__
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+
+      ~DbtTestOperator.__init__
+      ~DbtTestOperator.add_global_flags
+      ~DbtTestOperator.build_and_run_cmd
+      ~DbtTestOperator.exception_handling
+      ~DbtTestOperator.execute
+      ~DbtTestOperator.get_dbt_path
+      ~DbtTestOperator.get_env
+      ~DbtTestOperator.on_kill
+      ~DbtTestOperator.run_command
+      ~DbtTestOperator.subprocess_hook
+
+
+
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~DbtTestOperator.template_fields
+      ~DbtTestOperator.ui_color
