+++ docs/reference/dbt-reference.rst
@@ -1,0 +1,30 @@
+dbt
+=====================
+
+Cosmos provides a set of classes and operators to interact with dbt.
+
+.. currentmodule:: cosmos.providers.dbt
+
+DAGs and Task Groups
+---------------------
+
+.. autosummary::
+    :toctree: _generated/
+    :caption: DAGs and Task Groups
+
+    DbtDag
+    DbtTaskGroup
+
+
+Operators
+---------------------
+
+.. autosummary::
+    :toctree: _generated/
+    :caption: Operators
+
+    DbtRunOperator
+    DbtTestOperator
+    DbtLSOperator
+    DbtSeedOperator
+    DbtRunOperationOperator
