+++ docs/reference/index.rst
@@ -1,0 +1,7 @@
+API Reference
+=============
+
+.. toctree::
+    :maxdepth: 3
+
+    dbt <dbt-reference>
